requirements.txt
discord.py>=2.3.2
Flask>=2.3.3
aiohttp>=3.8.5